case-studies/ai-security-luxury/requirements.txt
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0